+++ exploration-module/src/main/java/io/edpn/backend/exploration/adapter/kafka/sender/KafkaMessageSender.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.exploration.application.dto.MessageDto;
 import io.edpn.backend.exploration.application.port.outgoing.topic.CreateTopicPort;
-import io.edpn.backend.exploration.application.port.outgoing.system.SendMessagePort;
+import io.edpn.backend.exploration.application.port.outgoing.message.SendMessagePort;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
