+++ exploration-module/src/main/java/io/edpn/backend/exploration/application/domain/SystemEliteIdRequest.java
@@ -1,0 +1,7 @@
+package io.edpn.backend.exploration.application.domain;
+
+import io.edpn.backend.util.Module;
+
+public record SystemEliteIdRequest(String systemName,
+                                   Module requestingModule) {
+}
