+++ exploration-module/src/main/java/io/edpn/backend/exploration/application/port/outgoing/message/SendMessagePort.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.exploration.application.port.outgoing.system;
+package io.edpn.backend.exploration.application.port.outgoing.message;
 
 import io.edpn.backend.exploration.application.dto.MessageDto;
