+++ exploration-module/src/test/java/io/edpn/backend/exploration/adapter/kafka/sender/KafkaMessageSenderTest.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.exploration.adapter.kafka.dto.KafkaMessageDto;
 import io.edpn.backend.exploration.application.dto.MessageDto;
+import io.edpn.backend.exploration.application.port.outgoing.message.SendMessagePort;
 import io.edpn.backend.exploration.application.port.outgoing.topic.CreateTopicPort;
-import io.edpn.backend.exploration.application.port.outgoing.system.SendMessagePort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
